backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit>=1.0.23",
    # Enable MCP support (LiveKit Agents raises an ImportError at runtime without the `mcp` extra)
    # so we can connect to Arcade's MCP Gateway over HTTP.
    "livekit-agents[mcp]~=1.3",
    "livekit-plugins-openai>=1.3.11",
    "livekit-plugins-assemblyai>=1.3.11",
    "livekit-plugins-elevenlabs>=1.3.11",
    "livekit-plugins-google>=1.3.11",
    "livekit-plugins-cartesia>=1.3.11",
    "livekit-plugins-silero>=1.3.11",
    "livekit-plugins-turn-detector>=1.3.11",  # Add as optional dependency
    "python-dotenv>=1.2.1",
    "supabase>=2.27.2",
    "livekit-plugins-noise-cancellation>=0.2.5",
    "python-dateutil>=2.9.0.post0",
    # Arcade.dev SDK (used by backend/read_gmail.py)
    "arcadepy>=0.1.0",
]

[dependency-groups]
dev = [
    "pyright>=1.1.408",
]
